email-bot/requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0
google-genai==1.10.0
python-dotenv==1.0.1
requests==2.31.0